pyproject.toml
[project]
name = "amazon-ads-campaigns"
version = "1.0.0"
description = "Enterprise-grade backend for Amazon Ads campaign management"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "dev@example.com"}
]
keywords = ["django", "amazon-ads", "campaigns", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15.0",
    "django-filter>=24.0",
    "drf-spectacular>=0.27.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "django-celery-beat>=2.6.0",
    "django-celery-results>=2.5.0",
    "python-decouple>=3.8",
    "dj-database-url>=2.1.0",
    "structlog>=24.0.0",
    "whitenoise>=6.6.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "factory-boy>=3.3.0",
    "faker>=28.0.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
    "pre-commit>=3.7.0",
    "ipython>=8.0.0",
]

production = [
    "gunicorn>=22.0.0",
    "psycopg2-binary>=2.9.0",
    "sentry-sdk>=2.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "config*", "integrations*"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
    "staticfiles",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "DJ",  # flake8-django
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config", "integrations"]

[tool.black]
line-length = 100
target-version = ["py311"]
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | migrations
    | staticfiles
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["apps", "integrations"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
show_missing = true
